# docs/swagger.yaml
basePath: /v1
definitions:
...
  dto.BorrowingResp:
    properties:
      borrow_date:
        type: string
      borrowed_book:
        type: string
      borrower_person:
        type: string
      id:
        type: integer
      return_date:
        type: string
    type: object
  dto.BorrowingUpdateReq:
    properties:
      book_id:
        type: integer
      borrow_date:
        type: string
      person_id:
        type: integer
      return_date:
        type: string
    required:
    - book_id
    - person_id
    type: object
  dto.ErrorResponse:
    properties:
      errors: {}
      message:
        type: string
      success:
        example: false
        type: boolean
    type: object
  dto.PersonCreateReq:
    properties:
      birth_date:
        type: string
      fullname:
        maxLength: 56
        minLength: 4
        type: string
      gender:
        enum:
        - m
        - f
        type: string
    required:
    - birth_date
    - fullname
    - gender
    type: object
  dto.PersonDetailResp:
    properties:
      age:
        type: integer
      fullname:
        type: string
      gender:
        type: string
      id:
        type: integer
    type: object
  dto.PersonUpdateReq:
    properties:
      birth_date:
        type: string
      fullname:
        maxLength: 56
        minLength: 4
        type: string
      gender:
        enum:
        - m
        - f
        type: string
    required:
    - birth_date
    - fullname
    - gender
    type: object
...
  dto.SuccessResponse-any:
    properties:
      data: {}
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.SuccessResponse-array_dto_BorrowingResp:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.BorrowingResp'
        type: array
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.SuccessResponse-array_dto_PersonDetailResp:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.PersonDetailResp'
        type: array
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.SuccessResponse-dto_BorrowingResp:
    properties:
      data:
        $ref: '#/definitions/dto.BorrowingResp'
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.SuccessResponse-dto_PersonDetailResp:
    properties:
      data:
        $ref: '#/definitions/dto.PersonDetailResp'
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:3000
info:
  contact:
    email: mark.p.e.muhammad@gmail.com
    name: Mark Muhammad
  description: This is a boilerplate project, please update accordingly.
  license:
    name: MIT
  title: Base API Service
  version: "1.0"
paths:
...
  /borrowings:
    get:
      description: Get a list of borrowings.
      parameters:
      - description: Borrowing's name
        in: query
        name: q
        type: string
      - description: Data offset
        in: query
        name: s
        type: integer
      - description: Data limit
        in: query
        name: l
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-array_dto_BorrowingResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a list of borrowings
    post:
      consumes:
      - application/json
      description: Create a borrowing.
      parameters:
      - description: Borrowing's detail
        in: body
        name: detail
        required: true
        schema:
          $ref: '#/definitions/dto.BorrowingUpdateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a borrowing
  /borrowings/{id}:
    delete:
      description: Delete a borrowing.
      parameters:
      - description: Borrowing's ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a borrowing
    get:
      description: Get a borrowing's detail.
      parameters:
      - description: Borrowing's ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_BorrowingResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a borrowing's detail
    put:
      consumes:
      - application/json
      description: Update a borrowing's detail.
      parameters:
      - description: Borrowing's ID
        in: path
        name: id
        required: true
        type: integer
      - description: Borrowing's detail
        in: body
        name: detail
        required: true
        schema:
          $ref: '#/definitions/dto.BorrowingUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a borrowing's detail
  /persons:
    get:
      description: Get a list of person.
      parameters:
      - description: Person's name
        in: query
        name: q
        type: string
      - description: Data offset
        in: query
        name: s
        type: integer
      - description: Data limit
        in: query
        name: l
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-array_dto_PersonDetailResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a list of person
    post:
      consumes:
      - application/json
      description: Create a person.
      parameters:
      - description: Person's detail
        in: body
        name: detail
        required: true
        schema:
          $ref: '#/definitions/dto.PersonCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a person
  /persons/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a Person entity by ID
      operationId: person-repository-delete
      parameters:
      - description: Person ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Person deleted successfully
          schema: {}
        "404":
          description: Person not found
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Delete a Person by ID
    get:
      description: Get a person's detail.
      parameters:
      - description: Person's ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_PersonDetailResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get a person's detail
    put:
      consumes:
      - application/json
      description: Update a person's detail.
      parameters:
      - description: Person's ID
        in: path
        name: id
        required: true
        type: integer
      - description: Person's detail
        in: body
        name: detail
        required: true
        schema:
          $ref: '#/definitions/dto.PersonUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a person's detail
...
securityDefinitions:
  BearerAuth:
    description: Bearer auth containing JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"